+++ app/resource/fields.py
@@ -1,7 +1,4 @@
-import unicodedata
-import re
-
-from app import bcrypt
+from app import bcrypt, utils
 from mongoengine import StringField
 
 
@@ -27,15 +24,10 @@ class PasswordField(StringField):
 
 class SlugField(StringField):
 
-    STRIP_REGEXP = re.compile(r'[^\w\s-]')
-    HYPHENATE_REGEXP = re.compile(r'[-\s]+')
-
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs, unique=True)
 
-    def slugify(self, value):
-        if not isinstance(value, str):
-            value = str(value)
-        value = unicodedata.normalize('NFKD', value).encode('ascii', 'ignore')
-        value = str(self.STRIP_REGEXP.sub('', value).strip().lower())
-        return self.HYPHENATE_REGEXP.sub('-', value)
+    def validate(self, value):
+        if not utils.SLUG_REGEX.match(value):
+            self.error("Provided string is not a slug.")
+        super().validate(value)
